% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_helpers.R
\name{bind_repression}
\alias{bind_repression}
\title{helper function to bind the repression vector to df_i in gdsc_repress}
\usage{
bind_repression(df, cell_lines, i, repression_vector, v_rm)
}
\arguments{
\item{df}{tidy dataframe containing expression data for a number of cell lines}

\item{cell_lines}{unique vector of cell lines}

\item{i}{index value for `cell_lines}

\item{repression_vector}{output of node_repression function in gdsc_repress}

\item{v_rm}{which node should we remove?}
}
\description{
helper function to bind the repression vector to df_i in gdsc_repress
}
