% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{calc_dnp_i}
\alias{calc_dnp_i}
\title{helper function to calculate dnp for one sample}
\usage{
calc_dnp_i(df, g)
}
\arguments{
\item{df}{dataframe with one cell line + log expression}

\item{g}{igraph object containing ppi info}
}
\value{
same dataframe with dnp calculated for each gene.
}
\description{
helper function to calculate dnp for one sample
}
