% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_helpers.R
\name{calc_exp}
\alias{calc_exp}
\title{helper function to grab the named expression vector for a given cell line}
\usage{
calc_exp(df, cell_lines, i)
}
\arguments{
\item{df}{tidy dataframe containing expression data for a number of cell lines}

\item{cell_lines}{unique vector of cell lines}

\item{i}{index value for `cell_lines}
}
\description{
helper function to grab the named expression vector for a given cell line
}
