% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_helpers.R
\name{calc_gdsc_np}
\alias{calc_gdsc_np}
\title{function to calculate network potential for every gene/cell line combo in the GDSC}
\usage{
calc_gdsc_np(cache = NULL, ncores = 1)
}
\arguments{
\item{cache}{}
}
\description{
function to calculate network potential for every gene/cell line combo in the GDSC
}
