% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_helpers.R
\name{calc_np_i}
\alias{calc_np_i}
\title{helper function to calculate np for one cell line in the gdsc}
\usage{
calc_np_i(df, g)
}
\arguments{
\item{df}{dataframe with one cell line + log expression}

\item{g}{igraph object containing ppi info}
}
\value{
same dataframe with np calculated for each gene.
}
\description{
helper function to calculate np for one cell line in the gdsc
}
