% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{clean_dna}
\alias{clean_dna}
\title{Function to clean dna variant data from gdsc
data can be fouund \href{https://www.cancerrxgene.org/gdsc1000/GDSC1000_WebResources/Home.html}{here}.}
\usage{
clean_dna(df)
}
\arguments{
\item{df}{unclean data frame}
}
\description{
Function to clean dna variant data from gdsc
data can be fouund \href{https://www.cancerrxgene.org/gdsc1000/GDSC1000_WebResources/Home.html}{here}.
}
