% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{clean_expression}
\alias{clean_expression}
\title{Function to clean expression data from gdsc
We'll read in the GDSC expression data, which comes from microarray experiments.
Details on the experimental protocols can be found
\href{https://www.cancerrxgene.org/gdsc1000/GDSC1000_WebResources/Home.html}{here}.}
\usage{
clean_expression(df)
}
\arguments{
\item{df}{unclean data frame}
}
\description{
Function to clean expression data from gdsc
We'll read in the GDSC expression data, which comes from microarray experiments.
Details on the experimental protocols can be found
\href{https://www.cancerrxgene.org/gdsc1000/GDSC1000_WebResources/Home.html}{here}.
}
