% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{clean_gdsc}
\alias{clean_gdsc}
\title{clean the GDSC data}
\usage{
clean_gdsc(cache = NULL)
}
\arguments{
\item{cache}{filepath specifying where to store store downloaded/ processed data}
}
\description{
clean the GDSC data
}
