% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{clean_meta}
\alias{clean_meta}
\title{helper function for processing the gdsc metadata}
\usage{
clean_meta(df)
}
\arguments{
\item{df}{unclean data frame}
}
\description{
Here we read in the meta data for GDSC. This contains all the cell lines by
various identifiers (we use \code{COSMIC_ID}) and descriptions of each line,
including tissue and histology of origin. We also do a quick cleaning step by
adjusting a variable name containing a forward slash (\code{/}) to avoid errors later.
We will also split out epithelial vs non epithelial cell lines.
}
