% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_helpers.R
\name{clean_output}
\alias{clean_output}
\title{function to summarise output from \code{GDSC_repress}}
\usage{
clean_output(df)
}
\arguments{
\item{df}{data frame containing NP difference from in-silico repression of a given gene

Computes the total network potential/ difference in network potential with repression of a given gene,
expression for the repressed gene, network potential for the repressed gene.}
}
\description{
function to summarise output from \code{GDSC_repress}
}
