% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{clean_response}
\alias{clean_response}
\title{Import and clean drug response data}
\usage{
clean_response(df)
}
\arguments{
\item{df}{unclean data frame}
}
\description{
Here we import the drug response data from the second version of the GDSC dataset (GDSC2).
Each cell line can still be identified using \code{COSMIC_ID}.
We clean the data by adjusting data types and converting IC50 from natural log to log2.
}
