% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_null.R
\name{combine_null}
\alias{combine_null}
\title{.combine function for compute_null foreach looping structure}
\usage{
combine_null(...)
}
\description{
.combine function for compute_null foreach looping structure
}
