% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{compute_np}
\alias{compute_np}
\title{main function to compute np from a user-provided expression matrix.}
\usage{
compute_np(cache = NULL)
}
\arguments{
\item{cache}{user-provided filepath for where to store data etc}
}
\value{
tidy data frame with one column for expression and another for np
}
\description{
main function to compute np from a user-provided expression matrix.
}
