% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{experiment_breakout}
\alias{experiment_breakout}
\title{helper function to split experiment names into constituent parts}
\usage{
experiment_breakout(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
this is highly specific  to the miR paper
}
