% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_filter.R
\name{filter_edges}
\alias{filter_edges}
\title{Function to reduce size of network by random culling of edges.}
\usage{
filter_edges()
}
\arguments{
\item{prop}{proportion of edges to remove}
}
\description{
Function to reduce size of network by random culling of edges.
}
