% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_filter.R
\name{filter_nodes}
\alias{filter_nodes}
\title{Function to reduce size of network based on either gene expression or network potential}
\usage{
filter_nodes()
}
\arguments{
\item{df}{tidy dataframe containing gene expression +/- network potential.}

\item{prop}{proportion of nodes to keep.}
}
\description{
Function to reduce size of network based on either gene expression or network potential
}
