% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_helpers.R
\name{GDSC_repress}
\alias{GDSC_repress}
\title{function to do in-silico repression for a given gene on all cell lines in the GDSC}
\usage{
GDSC_repress(cache = NULL, v_rm = "EGFR", ncores = 1)
}
\arguments{
\item{cache}{user-provided filepath for where to store data etc}

\item{v_rm}{which node should we remove}

\item{ncores}{number of cores to use for calculations}
}
\description{
function to do in-silico repression for a given gene on all cell lines in the GDSC
}
