% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{get_dr}
\alias{get_dr}
\title{helper function to get drug response data}
\usage{
get_dr(cache)
}
\arguments{
\item{cache}{filepath specifying where to store store downloaded/ processed data}
}
\description{
helper function to get drug response data
}
