% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{get_exp}
\alias{get_exp}
\title{helper function to download expression data specifically}
\usage{
get_exp(cache)
}
\arguments{
\item{cache}{filepath specifying where to store store downloaded/ processed data}
}
\description{
helper function to download expression data specifically
}
