% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{get_gdsc}
\alias{get_gdsc}
\title{Download GDSC Data}
\usage{
get_gdsc(cache = NULL)
}
\arguments{
\item{cache}{filepath specifying where to store store downloaded/ processed data}
}
\description{
Function to download the latest version of RNA expression, drug response,
and metadata from the Genomics of Drug Sensitivity in Cancer database.
link \href{https://www.cancerrxgene.org}{here}
}
