% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingest.R
\name{get_meta}
\alias{get_meta}
\title{helper function to get the cell line metadata}
\usage{
get_meta(cache)
}
\arguments{
\item{cache}{filepath specifying where to store store downloaded/ processed data}
}
\description{
helper function to get the cell line metadata
}
