% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_null.R
\name{get_topn}
\alias{get_topn}
\title{Helper function for compute_null - returns the top n genes by dnp for each sample}
\usage{
get_topn(df, n_genes)
}
\description{
Helper function for compute_null - returns the top n genes by dnp for each sample
}
