% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{load_ppi}
\alias{load_ppi}
\title{load ppi using crosstalkr imported functions}
\usage{
load_ppi(cache, min_score, ppi)
}
\arguments{
\item{cache}{user-provided filepath for where to store data etc}

\item{min_score}{if ppi is stringdb, which mininum score should we use to filter edges?}

\item{ppi}{should we use biogrid or stringdb for the PPI}
}
\description{
load ppi using crosstalkr imported functions
}
