% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{null_np}
\alias{null_np}
\title{Function to generate a null distribution of change in network potential for a given set of genes}
\usage{
null_np(g, exp, v_rm)
}
\arguments{
\item{g}{igraph network object}

\item{v_rm}{index of vertices to remove}
}
\description{
This is accomplished by randomly permuting the nodes in a PPI n times and re-calculating the change in network potential for each permutation
}
