% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{tidy_expression}
\alias{tidy_expression}
\title{helper function to convert expression matrix to tidy dataframe (if not already)}
\usage{
tidy_expression(df)
}
\description{
helper function to convert expression matrix to tidy dataframe (if not already)
}
